switched decimal to double in submodel, added km/h for wind speed

+++ DrachenwetterLambda/Models/Submodels/MainPrediction.cs
@@ -2,6 +2,6 @@
 {
     public struct MainPrediction
     {
-        public decimal Temp { get; set; }
+        public double Temp { get; set; }
     }
 }
